Move valve entity classes to entity module
